requirements.txt
Django
djangorestframework
requests
psycopg2
python-dotenv
